removing unused TS files
we went on an odyssey with ChatGPT and Claude to make external TS files work but Construct 3 doesn't seem to support that without running a converter to JS, which we don't want to do, so we moved all the TS to inline in the event sheets. the scripts folder is really just storage for the code that's now in the event sheets.